re-factor to remove require cycles
