+++ app/components/index.ts
@@ -12,7 +12,6 @@ export * from "./Avatar"
 export * from "./ButtonLink"
 export * from "./Tag"
 export * from "./IconButton"
-export * from "./CarouselCard"
 export * from "./BoxShadow"
-export * from "./Carousel"
+export * from "./carousel/Carousel"
 export * from "./ScrollToButton"
