+++ app/services/api/react-query.ts
@@ -1,0 +1,4 @@
+import { QueryClient } from "@tanstack/react-query"
+
+// Creating a react-query client with the stale time set to "Infinity" so that its never stale
+export const queryClient = new QueryClient({ defaultOptions: { queries: { staleTime: Infinity } } })
